TEST: Copy test data folder and cd to temp path during interface doctests
